tools/mlflow/requirements.txt
mlflow
awscli
mysqlclient
mysql-connector
boto3